Common functions moved to utils.lua. Indentation ...
